Move savefile creation into savefileservice
