+++ DragaliaAPI.Database.Test/DragaliaAPI.Database.Test.csproj
@@ -27,6 +27,7 @@
 	<ItemGroup>
 		<ProjectReference Include="..\DragaliaAPI.Database\DragaliaAPI.Database.csproj" />
 		<ProjectReference Include="..\DragaliaAPI.Shared\DragaliaAPI.Shared.csproj" />
+		<ProjectReference Include="..\DragaliaAPI\DragaliaAPI.csproj" />
 	</ItemGroup>
 
 </Project>
