+++ DragaliaAPI.Database.Test/Entities/DbPlayerCharaDataTest.cs
@@ -16,10 +16,7 @@ public class DbPlayerCharaDataTest
     [Fact]
     public void ManaCirclePieceIdList_Set_SetsExpectedValue()
     {
-        DbPlayerCharaData chara = DbPlayerCharaDataFactory.Create(
-            "id",
-            MasterAsset.CharaData.Get(Charas.GalaNedrick)
-        );
+        DbPlayerCharaData chara = new("id", Charas.GalaNedrick);
 
         SortedSet<int> input = new(Enumerable.Range(1, 50));
