+++ DragaliaAPI.Database/Entities/DbPlayerCharaData.cs
@@ -1,7 +1,12 @@
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using DragaliaAPI.Shared.Definitions.Enums;
+using DragaliaAPI.Shared.MasterAsset;
+using DragaliaAPI.Shared.MasterAsset.Models;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace DragaliaAPI.Database.Entities;
 
@@ -10,18 +15,14 @@ public class DbPlayerCharaData : IDbHasAccountId
 {
     /// <inheritdoc/>
     [Column("DeviceAccountId")]
-    [Required]
-    [ForeignKey("DbDeviceAccount")]
-    public string DeviceAccountId { get; set; } = null!;
+    public required string DeviceAccountId { get; set; }
 
     [Column("CharaId")]
-    [Required]
     [TypeConverter(typeof(EnumConverter))]
-    public Charas CharaId { get; set; }
+    public required Charas CharaId { get; set; }
 
     [Column("Rarity")]
-    [Required]
-    public byte Rarity { get; set; }
+    public required byte Rarity { get; set; }
 
     [Column("Exp")]
     public int Exp { get; set; } = 0;
@@ -53,41 +54,37 @@ public class DbPlayerCharaData : IDbHasAccountId
     public byte Skill1Level { get; set; } = 1;
 
     [Column("Skill2Lvl")]
-    public byte Skill2Level { get; set; } = 1;
+    public byte Skill2Level { get; set; } = 0;
 
     [Column("Abil1Lvl")]
     public byte Ability1Level { get; set; } = 1;
 
     [Column("Abil2Lvl")]
-    public byte Ability2Level { get; set; } = 1;
+    public byte Ability2Level { get; set; }
 
     [Column("Abil3Lvl")]
-    public byte Ability3Level { get; set; } = 1;
+    public byte Ability3Level { get; set; }
 
     [Column("BurstAtkLvl")]
-    public byte BurstAttackLevel { get; set; } = 1;
+    public byte BurstAttackLevel { get; set; }
 
     [Column("ComboBuildupCount")]
-    public int ComboBuildupCount { get; set; } = 0;
+    public int ComboBuildupCount { get; set; }
 
     [Column("HpBase")]
-    [Required]
-    public ushort HpBase { get; set; }
+    public required ushort HpBase { get; set; }
 
     [Column("HpNode")]
-    [Required]
-    public ushort HpNode { get; set; }
+    public ushort HpNode { get; set; } = 0;
 
     [NotMapped]
     public int Hp => this.HpBase + this.HpNode;
 
     [Column("AtkBase")]
-    [Required]
-    public ushort AttackBase { get; set; }
+    public required ushort AttackBase { get; set; }
 
     [Column("AtkNode")]
-    [Required]
-    public ushort AttackNode { get; set; }
+    public ushort AttackNode { get; set; } = 0;
 
     [NotMapped]
     public int Attack
@@ -96,35 +93,25 @@ public class DbPlayerCharaData : IDbHasAccountId
     }
 
     [Column("ExAbility1Lvl")]
-    [Required]
     public byte ExAbilityLevel { get; set; }
 
     [Column("ExAbility2Lvl")]
-    [Required]
     public byte ExAbility2Level { get; set; }
 
     [Column("IsTemp")]
-    [Required]
-    [TypeConverter(typeof(BooleanConverter))]
     public bool IsTemporary { get; set; }
 
     [Column("IsUnlockEditSkill")]
-    [Required]
     public bool IsUnlockEditSkill { get; set; }
 
     [Column("ManaNodeUnlockCount")]
-    [Required]
     public ushort ManaNodeUnlockCount { get; private set; }
 
     [Column("ListViewFlag")]
-    [Required]
-    [TypeConverter(typeof(BooleanConverter))]
     public bool ListViewFlag { get; set; }
 
     [Column("GetTime")]
-    [Required]
-    [TypeConverter(typeof(DateTimeOffsetConverter))]
-    public DateTimeOffset GetTime { get; set; }
+    public DateTimeOffset GetTime { get; set; } = DateTime.UtcNow;
 
     [NotMapped]
     public SortedSet<int> ManaCirclePieceIdList
@@ -133,5 +120,48 @@ public class DbPlayerCharaData : IDbHasAccountId
         set =>
             ManaNodeUnlockCount = (ushort)
                 ManaNodesUtil.SetManaCircleNodesFromSet(value, (ManaNodes)ManaNodeUnlockCount);
+    }
+
+    /// <summary>
+    /// EF Core / test constructor.
+    /// </summary>
+    public DbPlayerCharaData() { }
+
+    /// <summary>
+    /// User-facing constructor.
+    /// </summary>
+    /// <param name="deviceAccountId">Primary key.</param>
+    [SetsRequiredMembers]
+    public DbPlayerCharaData(string deviceAccountId, Charas id)
+    {
+        CharaData data = MasterAsset.CharaData.Get(id);
+
+        byte rarity = (byte)data.Rarity;
+        ushort rarityHp;
+        ushort rarityAtk;
+
+        switch (rarity)
+        {
+            case 3:
+                rarityHp = (ushort)data.MinHp3;
+                rarityAtk = (ushort)data.MinAtk3;
+                break;
+            case 4:
+                rarityHp = (ushort)data.MinHp4;
+                rarityAtk = (ushort)data.MinAtk4;
+                break;
+            case 5:
+                rarityHp = (ushort)data.MinHp5;
+                rarityAtk = (ushort)data.MinAtk5;
+                break;
+            default:
+                throw new UnreachableException("Invalid rarity!");
+        }
+
+        this.DeviceAccountId = deviceAccountId;
+        this.Rarity = rarity;
+        this.CharaId = id;
+        this.HpBase = rarityHp;
+        this.AttackBase = rarityAtk;
     }
 }
