+++ DragaliaAPI.Database/Repositories/IDeviceAccountRepository.cs
@@ -2,12 +2,10 @@
 
 namespace DragaliaAPI.Database.Repositories;
 
+[Obsolete("Used by pre-BaaS login flow")]
 public interface IDeviceAccountRepository : IBaseRepository
 {
     Task AddNewDeviceAccount(string id, string hashedPassword);
 
     Task<DbDeviceAccount?> GetDeviceAccountById(string id);
-
-    Task CreateNewSavefile(string deviceAccountId);
-    Task CreateNewSavefileBase(string deviceAccountId);
 }
