+++ DragaliaAPI/Services/ISavefileService.cs
@@ -4,6 +4,8 @@ namespace DragaliaAPI.Services;
 
 public interface ISavefileService
 {
+    Task CreateNewSavefile(string deviceAccountId);
+    Task CreateNewSavefileBase(string deviceAccountId);
     Task Import(string deviceAccountId, LoadIndexData savefile);
     Task Reset(string deviceAccountId);
 }
